Add del and update methods

+++ src/main/java/com/main/uatouristassistant/repository/CommentAndRatingRepository.java
@@ -3,7 +3,7 @@ package com.main.uatouristassistant.repository;
 import com.main.uatouristassistant.entity.CommentAndRating;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentAndRatingRepository extends JpaRepository<CommentAndRating,Integer> {
+public interface CommentAndRatingRepository extends JpaRepository<CommentAndRating,Long> {
     CommentAndRating findCommentAndRateByUserId(Long id);
     CommentAndRating findCommentAndRateByPlaceId(Long id);
 }
